+++ tests/test_extension/roots/test-translations/conf.py
@@ -1,0 +1,9 @@
+project = 'Multi-languages docs-versions-menu test'
+
+extensions = [
+    'docs_versions_menu',
+]
+
+docs_versions_menu_conf = {
+    'url_version_scheme': 'translations',
+}